+++ src/main/java/project/flipnote/like/model/response/LikeTargetResponse.java
@@ -1,4 +1,4 @@
-package project.flipnote.like.model;
+package project.flipnote.like.model.response;
 
 public abstract class LikeTargetResponse {
 	public abstract Long getId();
