+++ src/main/java/project/flipnote/like/service/LikeTargetFetchService.java
@@ -13,7 +13,7 @@ import lombok.RequiredArgsConstructor;
 import project.flipnote.common.exception.BizException;
 import project.flipnote.like.entity.LikeTargetType;
 import project.flipnote.like.exception.LikeErrorCode;
-import project.flipnote.like.model.LikeTargetResponse;
+import project.flipnote.like.model.response.LikeTargetResponse;
 import project.flipnote.like.service.fetcher.LikeTargetFetcher;
 
 @RequiredArgsConstructor
