+++ src/main/java/project/flipnote/like/service/fetcher/LikeTargetFetcher.java
@@ -4,7 +4,7 @@ import java.util.Map;
 import java.util.Set;
 
 import project.flipnote.like.entity.LikeTargetType;
-import project.flipnote.like.model.LikeTargetResponse;
+import project.flipnote.like.model.response.LikeTargetResponse;
 
 public interface LikeTargetFetcher<T extends LikeTargetResponse> {
 	LikeTargetType getTargetType();
